perf(transactions): renames create transaction methods and avoids unnecessary calls by using the com
When using commit=1 the transaction is created and committed in a single step server-side, no two
calls necessary. As a result createOrCommitTransaction has been updated to use this method for
performance reasons. The previous createTransaction call has been renamed to prepareTransaction and
createOrCommitTransaction is now named simply createTransaction.

closes #3 BREAKING CHANGE: createTransaction renamed to prepareTransaction,
createAndCommitTransaction renamed to createTransaction