fix(otp): adds better two factor auth support and docs
Previously the SDK relied upon the Unauthorized error message being returned to detect whether a
OTP was required when creating a PAT however the unauthorized error can also be returned if the
username/password is incorrect so that wasnt good enough. It now checks for the x-bitreserve-otp
header and responds with an object containing an otp property instead if two factor authentication
is enabled.